adding more information to the output so fab info is decipherable

+++ corehq/apps/hqadmin/management/commands/clear_supervisor_confs.py
@@ -3,6 +3,7 @@ from optparse import make_option
 
 from django.core.management.base import BaseCommand
 from django.conf import settings
+import sys
 
 
 class Command(BaseCommand):
@@ -16,8 +17,11 @@ class Command(BaseCommand):
     def handle(self, *args, **options):
         conf_dir = options['conf_location']
         environment = settings.SERVER_ENVIRONMENT
+        if not os.path.exists(conf_dir):
+            sys.exit("[clear_supervisor_confs] Error: the path %s is not reachable by this process" % conf_dir)
         files = os.listdir(conf_dir)
         env_confs = filter(lambda x: x.startswith('%s_' % environment) and x.endswith('.conf'), files)
         for c in env_confs:
             os.remove(os.path.join(conf_dir, c))
+            print "\t[clear_supervisor_confs] Removed supervisor configuration file: %s" % c
